+++ src/strcpy.rs
@@ -1,6 +1,5 @@
 //! Rust implementation of C library function `strcpy`
 //!
-//! Copyright (c) Dion Dokter 2022
 //! Licensed under the Blue Oak Model Licence 1.0.0
 
 use crate::CChar;
@@ -9,14 +8,13 @@ use crate::CChar;
 /// (core::ptr::null()) gives undefined behaviour.
 #[no_mangle]
 pub unsafe extern "C" fn strcpy(dest: *mut CChar, src: *const CChar) -> *const CChar {
-	let mut i = 0isize;
+	let mut i = 0;
 	loop {
 		*dest.offset(i) = *src.offset(i);
-
-		if *src.offset(i) == b'\0' {
+		let c = *dest.offset(i);
+		if c == 0 {
 			break;
 		}
-
 		i += 1;
 	}
 	dest
@@ -32,16 +30,16 @@ mod test {
 		let mut dest = *b"abcdef"; // no null terminator
 		let result = unsafe { strcpy(dest.as_mut_ptr(), src.as_ptr()) };
 		assert_eq!(
-			unsafe { core::slice::from_raw_parts(result, 6) },
-			*b"hi\0def"
+			unsafe { core::slice::from_raw_parts(result, 5) },
+			*b"hi\0de"
 		);
 	}
 
 	#[test]
 	fn two() {
 		let src = b"hi\0";
-		let mut dest = [0u8; 3];
+		let mut dest = [0u8; 2]; // no space for null terminator
 		let result = unsafe { strcpy(dest.as_mut_ptr(), src.as_ptr()) };
-		assert_eq!(unsafe { core::slice::from_raw_parts(result, 3) }, b"hi\0");
+		assert_eq!(unsafe { core::slice::from_raw_parts(result, 2) }, b"hi");
 	}
 }